event

+++ app/src/main/java/com/example/micke/labb2mobil/Model/GameObjects/Position.java
@@ -44,7 +44,7 @@ public class Position implements Serializable {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
 
@@ -52,7 +52,7 @@ public class Position implements Serializable {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
